[Refactor] Move vintageClipboard to utils/
