+++ src/scripts/app.ts
@@ -56,7 +56,7 @@ import { shallowReactive } from 'vue'
 import { type IBaseWidget } from '@comfyorg/litegraph/dist/types/widgets'
 import { useWorkflowService } from '@/services/workflowService'
 import { useWidgetStore } from '@/stores/widgetStore'
-import { deserialiseAndCreate } from '@/extensions/core/vintageClipboard'
+import { deserialiseAndCreate } from '@/utils/vintageClipboard'
 import { st } from '@/i18n'
 import { normalizeI18nKey } from '@/utils/formatUtil'
 import { useExtensionService } from '@/services/extensionService'
