+++ src/utils/vintageClipboard.ts
